Saved recipes

+++ app/api/recipe_routes.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, request, session
-from ..models import Recipe, Category, db, Quantity, Step, Rating, Review
+from ..models import Recipe, Category, db, Quantity, Step, Rating, Review, User
 from ..forms import RecipeForm, QuantityForm, StepForm, ReviewForm, RatingForm
 from flask_login import login_required
 from app.aws import (upload_file_to_s3, get_unique_filename)
@@ -308,6 +308,10 @@ def delete_recipe(recipeId):
     quantities = Quantity.query.filter(Quantity.recipe_id == recipeId).all()
     steps = Step.query.filter(Step.recipe_id == recipeId).all()
 
+
+    if not recipe:
+        return {"error": "resource not found"}, 404
+
     if recipe and recipe.owner_id == int(session['_user_id']):
         db.session.delete(recipe)
         [db.session.delete(quantity) for quantity in quantities]
@@ -315,9 +319,6 @@ def delete_recipe(recipeId):
 
         db.session.commit()
         return {"message": "successful"}
-
-    elif not recipe:
-        return {"error": "resource not found"}, 404
 
     else:
         return {"error": "Unauthorized"}, 403
@@ -378,3 +379,37 @@ def post_rating(recipeId):
 
     else:
         return {"errors": form.errors}, 400
+
+
+@recipe.route('/<int:recipeId>/save', methods=['POST'])
+def save_recipe(recipeId):
+    """
+    Route that saves a recipe to the user's saved recipes
+    """
+
+    userId = int(session['_user_id'])
+
+    user = User.query.get(userId)
+    recipe = Recipe.query.get(recipeId)
+
+    user.saved_recipes.append(recipe)
+    db.session.commit()
+
+    return user.to_dict()
+
+
+@recipe.route('/<int:recipeId>/unsave', methods=['DELETE'])
+def unsave_recipe(recipeId):
+    """
+    Route that removes a recipe from the user's saved recipes
+    """
+
+    userId = int(session['_user_id'])
+
+    user = User.query.get(userId)
+    recipe = Recipe.query.get(recipeId)
+
+    user.saved_recipes.remove(recipe)
+    db.session.commit()
+
+    return user.to_dict()
